requirements.txt
Django==4.2
djangorestframework==3.14.0
django-filter==23.2
pyyaml==6.0.1
requests